+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 ----
 * remove deprecated ``memoryfile`` usage for ``write_raster_window()``
 * fix ``s3`` path for ``mapchete index``
+* add ``snap_bounds``, ``clip_bounds`` functions & ``effective_bounds`` to config
 
 ----
 0.22